Refactor room methods

+++ pkg/service/room.go
@@ -1,8 +1,7 @@
 package service
 
 import (
-	"errors"
-
+	. "github.com/architectv/property-task/pkg/error"
 	"github.com/architectv/property-task/pkg/model"
 	"github.com/architectv/property-task/pkg/repository"
 )
@@ -17,18 +16,38 @@ func NewRoomService(repo repository.Room) *RoomService {
 
 func (s *RoomService) Create(room *model.Room) (int, error) {
 	if room.Description == "" {
-		return 0, errors.New("description should not be empty")
+		return 0, ErrEmptyDescription
 	}
 	if room.Price <= 0 {
-		return 0, errors.New("price should be positive number")
+		return 0, ErrNotPositivePrice
 	}
+
 	return s.repo.Create(room)
 }
 
 func (s *RoomService) Delete(id int) error {
+	if id <= 0 {
+		return ErrNotPositiveId
+	}
 	return s.repo.Delete(id)
 }
 
-func (s *RoomService) GetAll(sortField string, asc bool) ([]*model.Room, error) {
-	return s.repo.GetAll(sortField, asc)
+func (s *RoomService) GetAll(sortField string) ([]*model.Room, error) {
+	desc := false
+
+	if sortField == "" {
+		sortField = "id"
+	} else if len(sortField) >= 2 {
+		if sortField[0] == '-' {
+			desc = true
+			sortField = sortField[1:]
+		}
+		if sortField != "id" && sortField != "price" {
+			return nil, ErrWrongSortField
+		}
+	} else {
+		return nil, ErrWrongSortField
+	}
+
+	return s.repo.GetAll(sortField, desc)
 }
